test/inspector/runtime/map-set-entry-preview-sideeffects-expected.txt

Checks that Map/Set entry preview does not trigger Proxy traps on prototype chain.

Running test: testMapSetEntryPreviewSideEffects
{
    gOPDCalled : 0
    ownKeysCalled : 0
}
